requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-extensions==2.2.8
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==1.5
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
